implements sharing

+++ client/src/components/ListItem.jsx
@@ -2,11 +2,11 @@ import { Link } from 'react-router-dom';
 import { format } from 'date-fns';
 import formatBytes from '../utils/formatBytes.js';
 import { useState } from 'react';
-import { SquarePen, Download, Trash } from 'lucide-react';
+import { SquarePen, Forward, Download, Trash } from 'lucide-react';
 import api from '../api/urls.js';
 import ConfirmBlock from './ConfirmBlock.jsx';
 import ItemEdit from './ItemEdit.jsx';
-
+import FolderShare from './FolderShare.jsx';
 export function ListItem({
   name,
   id,
@@ -16,10 +16,13 @@ export function ListItem({
   to,
   removeItem,
   updateItem,
+  shareItem,
+  shareToken,
 }) {
   const [remove, setRemove] = useState(false);
   const [edit, setEdit] = useState(false);
-  const isFolder = to.includes('folder');
+  const [share, setShare] = useState(false);
+  const isFolder = to?.includes('folder') || false;
   const handleRemove = () => {
     setRemove(false);
     removeItem();
@@ -46,22 +49,51 @@ export function ListItem({
         }}
       />
     );
+  if (share)
+    return (
+      <FolderShare
+        onCancelSave={(data) => {
+          setShare(false);
+          if (data) {
+            shareItem(data);
+          }
+        }}
+      />
+    );
+  const linkContent = (
+    <>
+      {' '}
+      <Icon className="text-primary dark:text-white shrink-0" />
+      <span className="truncate " title={name}>
+        {name}
+      </span>
+    </>
+  );
   const content = (
     <div className="flex  flex-col bg-white dark:bg-primaryDark justify-between border border-gray-400 p-1 mb-0.5">
       <div className="flex  ">
-        <Link
-          to={to}
-          className="flex min-w-0 basis-2/3 gap-1 clickable items-center"
-        >
-          <Icon className="text-primary shrink-0" />
-          <span className="truncate " title={name}>
-            {name}
-          </span>
-        </Link>
-        <div className="basis-1/6 text-center mx-5  ml-4">
+        {shareToken ? (
+          <a
+            href={to}
+            download
+            className="flex min-w-0 w-1/2 gap-1 clickable items-center"
+          >
+            {linkContent}
+          </a>
+        ) : (
+          <Link
+            to={shareToken ? '' : to}
+            target={shareToken ? '_blank' : ''}
+            rel={shareToken ? 'noopener noreferrer' : ''}
+            className="flex min-w-0 w-1/2 gap-1 clickable items-center"
+          >
+            {linkContent}
+          </Link>
+        )}
+        <div className="w-1/6 text-center   ">
           {size ? formatBytes(size) : ''}
         </div>
-        <div className="basis-1/6  whitespace-nowrap">
+        <div className="w-1/3  whitespace-nowrap text-right">
           {createdAt ? format(new Date(createdAt), 'PP, HH:mm') : ''}
         </div>
       </div>
@@ -83,9 +115,17 @@ export function ListItem({
           </span>
         )}
         {!isFolder && (
-          <a href={api.fileDownload(id)} download>
+          <a href={shareToken ? to : api.fileDownload(id)} download>
             <Download className="dark:stroke-white clickable text-primary" />
           </a>
+        )}
+        {shareItem && (
+          <span title="Delete" className="clickable">
+            <Forward
+              className="dark:stroke-white clickable text-primary"
+              onClick={() => setShare(true)}
+            />
+          </span>
         )}
       </div>
     </div>
